updated lat long types

+++ data/trip.go
@@ -20,8 +20,8 @@ type Trip struct {
 	End_date		string
 	Place_id		string
 	Photo_uri		string
-	Latitude		int
-	Longitude		int
+	Latitude		float64
+	Longitude		float64
 }
 
 type TripRepository struct {
@@ -56,8 +56,8 @@ func (r *TripRepository) FindTrips(ctx *gin.Context, userId int) ([]Trip, error)
 			End_date:	row.EndDate.Time.Format("2006-01-02"),
 			Place_id: 	row.PlaceID.String,
 			Photo_uri: 	row.PhotoUri.String,
-			Latitude:	int(row.Latitude.Int32),
-			Longitude:	int(row.Longitude.Int32),
+			Latitude:	float64(row.Latitude.Float64),
+			Longitude:	float64(row.Longitude.Float64),
 		}
 		trips = append(trips, trip)
 	}
@@ -77,8 +77,8 @@ func (r * TripRepository) CreateTrip(ctx *gin.Context, newTrip Trip) (error) {
 		EndDate:  sql.NullTime{Time: formattedEndDate, Valid: true},
 		PlaceID: sql.NullString{String: newTrip.Place_id, Valid: newTrip.Place_id != ""},
 		PhotoUri: sql.NullString{String: newTrip.Photo_uri, Valid: newTrip.Photo_uri != ""},
-		Latitude:	sql.NullInt32{Int32: int32(newTrip.Latitude), Valid: true},
-		Longitude:	sql.NullInt32{Int32: int32(newTrip.Longitude), Valid: true},
+		Latitude: sql.NullFloat64{Float64: newTrip.Latitude, Valid: true},
+		Longitude: sql.NullFloat64{Float64: newTrip.Longitude, Valid: true},
 	}
 
 	err := r.queries.CreateTrip(ctx, params);
